javascript template readme and polish menus
