+++ javascript/src/main.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import "./index.css";
+import "./main.css";
 import App from "./view/App";
 
 ReactDOM.render(
